+++ NutTray/parseconf.cpp
@@ -102,7 +102,7 @@ static void pconf_fatal(PCONF_CTX *ctx, const char *errtxt)
 
 static void add_arg_word(PCONF_CTX *ctx)
 {
-	int	argpos;
+	size_t	argpos;
 	size_t	wbuflen;
 
 	/* this is where the new value goes */
